Initial move of visualization to gui_tools (and all imports renamed).
