+++ stellarphot/gui_tools/fits_opener.py
